+++ package.json
@@ -29,6 +29,7 @@
     "jest-watch-typeahead": "0.6.0",
     "prettier": "2.0.5",
     "ts-jest": "25.5.1",
+    "type-fest": "^0.13.1",
     "typescript": "3.8.3"
   },
   "jest": {
