+++ src/index.ts
@@ -1,3 +1,4 @@
 export { demo } from './demo'
 export { Logger } from './logger'
-export { create, RootLogger, SettingsData, SettingsInput } from './root-logger'
+export { create, RootLogger } from './root-logger'
+export { Data as SettingsData, Input as SettingsInput } from './settings'
